import package name changes and location fixes
