+++ conductor-user/src/main/java/io/appform/conductor/user/auth/AuthData.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2023 Santanu Sinha
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appform.conductor.user.auth;
+
+import lombok.Value;
+
+/**
+ * Extracted auth data to be used by {@link ConductorAuthenticator}
+ * and {@link ConductorAuthorizer}
+ */
+@Value
+public class AuthData {
+    AuthSource source;
+    String token;
+}
