+++ conductor-user/src/main/java/io/appform/conductor/user/auth/ConductorAuthenticator.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2023 Santanu Sinha
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appform.conductor.user.auth;
+
+import io.appform.conductor.core.auth.ConductorUser;
+import io.appform.conductor.user.internalmodels.auth.UserTokenAuthData;
+import io.dropwizard.auth.AuthenticationException;
+import io.dropwizard.auth.Authenticator;
+import lombok.RequiredArgsConstructor;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import java.util.Optional;
+
+/**
+ *
+ */
+@Singleton
+@RequiredArgsConstructor(onConstructor_ = {@Inject})
+public class ConductorAuthenticator implements Authenticator<String, ConductorUser> {
+    private final UserAuthValidator userAuthValidator;
+
+
+    @Override
+    public Optional<ConductorUser> authenticate(String credentials) throws AuthenticationException {
+        return userAuthValidator.authenticatedSession(new UserTokenAuthData(credentials))
+                .map(ConductorUser::new);
+    }
+}
