+++ conductor-user/src/main/java/io/appform/conductor/user/internalmodels/auth/PasswordAuthData.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (c) 2023 Santanu Sinha
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appform.conductor.user.internalmodels.auth;
+
+import io.appform.conductor.model.usermgmt.UserAuthType;
+import lombok.EqualsAndHashCode;
+import lombok.ToString;
+import lombok.Value;
+
+/**
+ *
+ */
+@Value
+@EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
+public class PasswordAuthData extends UserAuthData {
+    String userId;
+    String email;
+    String password;
+
+    public PasswordAuthData(String userId, String email, String password) {
+        super(UserAuthType.PASSWORD);
+        this.userId = userId;
+        this.email = email;
+        this.password = password;
+    }
+
+    @Override
+    public <T> T accept(UserAuthDataVisitor<T> visitor) {
+        return visitor.visit(this);
+    }
+}
