+++ conductor-user/src/main/java/io/appform/conductor/user/internalmodels/auth/UserAuthData.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2023 Santanu Sinha
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appform.conductor.user.internalmodels.auth;
+
+import io.appform.conductor.model.usermgmt.UserAuthType;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+/**
+ *
+ */
+@Data
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public abstract class UserAuthData {
+    protected final UserAuthType type;
+
+    public abstract <T> T accept(final UserAuthDataVisitor<T> visitor);
+}
