+++ conductor-user/src/main/java/io/appform/conductor/user/internalmodels/auth/UserPasswordAuthDetails.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2023 Santanu Sinha
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.appform.conductor.user.internalmodels.auth;
+
+import lombok.AllArgsConstructor;
+import lombok.Value;
+import lombok.With;
+
+import java.util.Date;
+
+/**
+ *
+ */
+@Value
+@AllArgsConstructor
+public class UserPasswordAuthDetails {
+    String userId;
+    @With
+    String password;
+    @With
+    int failedPasswordAttempts;
+    Date created;
+    Date updated;
+}
