--- conductor-user/src/main/java/io/appform/conductor/user/usermanagement/UserLifecycleManager.java
@@ -1,494 +1,0 @@
-/*
- * Copyright (c) 2023 Santanu Sinha
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.appform.conductor.server.usermanagement;
-
-import at.favre.lib.crypto.bcrypt.BCrypt;
-import com.google.common.base.Strings;
-import io.appform.conductor.model.attributes.AttributeScopeType;
-import io.appform.conductor.model.auth.Role;
-import io.appform.conductor.model.skills.SkillDefinition;
-import io.appform.conductor.model.skills.SkillValue;
-import io.appform.conductor.model.usermgmt.*;
-import io.appform.conductor.server.attributes.values.AttributeManager;
-import io.appform.conductor.server.auth.RoleStore;
-import io.appform.conductor.server.auth.UserAuthValidator;
-import io.appform.conductor.server.auth.UserRoleMappingStore;
-import io.appform.conductor.server.comms.MailSender;
-import io.appform.conductor.server.id.IdUtils;
-import io.appform.conductor.server.internalmodels.auth.PasswordAuthData;
-import io.appform.conductor.server.internalmodels.auth.UserTokenAuthData;
-import io.appform.conductor.server.skillmanagement.SkillStore;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import lombok.val;
-
-import javax.inject.Inject;
-import javax.inject.Provider;
-import java.util.*;
-import java.util.stream.Collectors;
-
-/**
- * User lifecycle manager
- */
-
-@Slf4j
-@RequiredArgsConstructor(onConstructor_ = {@Inject})
-public class UserLifecycleManager {
-    /**
-     * Default validity for {@link UserActivationToken}
-     */
-    private static final long DEFAULT_TOKEN_VALIDITY_MS = 86400_000; // 1 day
-
-    /**
-     * Default validity for {@link UserSession}
-     */
-    private static final long DEFAULT_SESSION_VALIDITY_MS = 7 * 86400_000; // 7 days
-
-    /**
-     * Once a user is in the following states, they cannot be reverted back or made actionable in any way
-     */
-
-
-    private final Provider<UserStore> userStore;
-    private final Provider<RoleStore> roleStore;
-    private final Provider<UserRoleMappingStore> roleMappingStore;
-    private final Provider<UserPasswordAuthStore> passwordAuthStore;
-    private final Provider<UserActivationTokenStore> userActivationTokenStore;
-    private final Provider<SessionStore> sessionStore;
-    private final UserAuthValidator userAuthValidator;
-    private final Provider<GroupStore> groupStore;
-    private final Provider<SkillStore> skillStore;
-    private final Provider<IdUtils> idUtils;
-    private final Provider<AttributeManager> attributeManager;
-    private final MailSender mailSender;
-
-    public Optional<User> userDetails(@NonNull final String userId) {
-        val roleId = roleMappingStore.get().roleForUser(userId).orElse(null);
-        return userStore.get().getById(userId)
-                .map(userSummary -> {
-                    val role = Strings.isNullOrEmpty(roleId)
-                               ? Optional.<Role>empty()
-                               : roleStore.get().read(roleId);
-                    val attributes = attributeManager.get()
-                            .read(AttributeScopeType.USER, userId)
-                            .stream()
-                            .filter(attribute -> Objects.nonNull(attribute.getDefinition()))
-                            .filter(attribute -> Objects.nonNull(attribute.getValue()))
-                            .collect(Collectors.toMap(attribute -> attribute.getDefinition().getName(),
-                                                      AttributeManager.MaterializedAttributeValue::getValue));
-                    return new User(userSummary,
-                                    role.orElse(null),
-                                    role.map(Role::getPermissions).orElse(Set.of()),
-                                    groupStore.get().findGroupsForUser(userSummary.getId()),
-                                    skillStore.get().listSkillsForUser(userId),
-                                    attributes);
-                });
-    }
-
-    /**
-     * This will create a user and send an activation link for the user.
-     * User will NOT be able to log in till activation is completed.
-     * The state for the user will be {@link UserState#CREATED}
-     *
-     * @param name     User full name
-     * @param email    User email id
-     * @param password Password for the user
-     * @return Materialized user details.
-     */
-    public Optional<UserSummary> createHumanUser(String name, String email, String password) {
-        val userDetails = userStore.get()
-                .create(idUtils.get().createUserInSameShard(email), name, UserType.HUMAN, email);
-        userDetails.ifPresent(user -> {
-            passwordAuthStore.get().set(user.getId(), hash(password));
-            createToken(user);
-        });
-        return userDetails;
-    }
-
-    public Optional<UserSummary> createSystemUser(String name, String email) {
-        val userDetails = userStore.get()
-                .create(idUtils.get().createUserInSameShard(email), name, UserType.SYSTEM, email)
-                .orElse(null);
-        if (null == userDetails) {
-            log.error("Could not create user: {}", name);
-            return Optional.empty();
-        }
-        return userStore.get().updateState(userDetails.getId(), UserState.ACTIVE);
-    }
-
-    public Optional<String> jwtForSession(String userId, String sessionId) {
-        return userStore.get()
-                .getById(userId)
-                .filter(userSummary -> userSummary.getState().equals(UserState.ACTIVE))
-                .flatMap(userSummary -> sessionStore.get()
-                        .getById(userId, sessionId)
-                        .map(session -> userAuthValidator.tokenFromSession(session, userSummary)));
-    }
-
-    public Optional<UserActivationToken> openToken(String userId) {
-        return userActivationTokenStore.get()
-                .getForUser(userId, Set.of(UserActivationTokenState.UNVALIDATED));
-    }
-
-    public Optional<UserSummary> showToken(String token) {
-        return userActivationTokenStore.get()
-                .getById(token)
-                .map(UserActivationToken::getUserId)
-                .flatMap(id -> userStore.get().getById(id));
-    }
-
-    /**
-     * This function will activate a user basis the provided token and the supplied password.
-     *
-     * @param token    Activation token generated by {@link #createHumanUser(String, String, String)}
-     * @param password Password entered by user
-     * @return Summary for a validated user
-     */
-    //TODO::EVENT
-    public Optional<UserSession> activateUser(String token, String password) {
-        if (Strings.isNullOrEmpty(token)) {
-            log.error("Empty token passed for validation");
-            return Optional.empty();
-        }
-        if (Strings.isNullOrEmpty(password)) {
-            log.error("Empty password for token: {}", token);
-            return Optional.empty();
-
-        }
-        val userToken = userActivationTokenStore.get().getById(token).orElse(null);
-        if (null == userToken) {
-            log.error("Invalid user activation token: {}", token);
-            return Optional.empty();
-        }
-        if (userToken.getState().equals(UserActivationTokenState.DELETED)) {
-            log.error("Token {} is already deleted", token);
-            return Optional.empty();
-        }
-        if (userToken.getState().equals(UserActivationTokenState.VALIDATED)) {
-            log.error("Token {} is already validated", token);
-            return Optional.empty();
-        }
-        val currentDate = new Date();
-        if (new Date(userToken.getCreated().getTime() + DEFAULT_TOKEN_VALIDITY_MS).before(currentDate)) {
-            log.error("Token {} is expired. Created date: {} Current Date: {}",
-                      token, userToken.getCreated(), currentDate);
-            return Optional.empty();
-        }
-
-        val userId = userToken.getUserId();
-        val userSession = userAuthValidator.authenticatedSession(new PasswordAuthData(userId, null, password)).orElse(
-                null);
-        if (null == userSession) {
-            log.error("No valid user for token {} userID {}", token, userId);
-            return Optional.empty();
-        }
-
-        val tokenUpdateStatus = userActivationTokenStore.get()
-                .updateTokenState(token, UserActivationTokenState.VALIDATED);
-        if (!tokenUpdateStatus) {
-            return Optional.empty();
-        }
-        log.info("Token {} successfully validated for user: {}", token, userId);
-        val userState = userSession.getUser().getSummary().getState();
-        if (userState.equals(UserState.CREATED)) {
-            val updatedDetails = userStore.get()
-                    .updateState(userId, UserState.ACTIVE)
-                    .orElse(null);
-            if (null != updatedDetails) {
-                return userDetails(userId)
-                        .map(user -> new UserSession(user,
-                                                     userSession.getSessionId(),
-                                                     userSession.getState(),
-                                                     userSession.getType(),
-                                                     userSession.getExpiry(),
-                                                     userSession.getJwt()));
-            }
-            else {
-                log.error("User state could not be updated to ACTIVE for {}", userId);
-                return Optional.empty();
-            }
-        }
-        log.warn("Token validated for user {} in state: {}", userId, userState);
-        return Optional.of(userSession);
-    }
-
-    /**
-     * Login a user and create a new session.
-     *
-     * @param email    Userid for the user
-     * @param password Password for the user
-     * @return A new user session
-     */
-    public Optional<UserSession> loginUser(String email, String password) {
-        //Get a validated user
-        return userAuthValidator.authenticatedSession(new PasswordAuthData(null, email, password));
-    }
-
-    /**
-     * Start a new session for a system user
-     *
-     * @param userId The user id for the system user
-     * @return A session for active system user or empty otherwise
-     */
-    public Optional<UserSession> startSystemUserSession(final String userId) {
-        return userStore.get()
-                .getById(userId)
-                .filter(user -> user.getType().equals(UserType.SYSTEM) && user.getState().equals(UserState.ACTIVE))
-                .flatMap(userAuthValidator::createSystemUserSession);
-    }
-
-    public boolean completeUserSession(final String userId, final String sessionId) {
-        return sessionStore.get().complete(userId, sessionId);
-    }
-
-    /**
-     * Logout the session
-     *
-     * @param session Session to be logged out of
-     * @return true if successful
-     */
-    public boolean logoutUser(UserSession session) {
-        //Get a validated user
-        return sessionStore.get()
-                .complete(session.getUser().getSummary().getId(), session.getSessionId());
-    }
-
-    /**
-     * Update name for the user
-     *
-     * @param userId user id for use whose name needs to be updated
-     * @param name   New name
-     * @return updated user
-     */
-    public Optional<UserSummary> updateUserName(final String userId, final String name) {
-        return userStore.get()
-                .update(userId,
-                        user -> user.withName(name));
-    }
-
-    public boolean changePassword(final String userId, final String oldPassword, final String newPassword) {
-        if (Strings.isNullOrEmpty(oldPassword) || Strings.isNullOrEmpty(newPassword)) {
-            log.error("Empty passwords: old {} new {}", oldPassword, newPassword);
-        }
-        else if (userAuthValidator.authenticate(new PasswordAuthData(userId, null, oldPassword))) {
-            log.info("Password updated for user: {}", userId);
-            return passwordAuthStore.get().updatePassword(userId, hash(newPassword)).isPresent();
-        }
-        return false;
-    }
-
-    /**
-     * This will validate the session for a user
-     *
-     * @param token The JWT token for a session
-     */
-    public Optional<UserSession> validateToken(String token) {
-        return userAuthValidator.authenticatedSession(new UserTokenAuthData(token));
-    }
-
-    /**
-     * Create a new {@link Group}
-     *
-     * @param name        name of the group
-     * @param description description of the group
-     * @param type
-     * @return The newly created group
-     */
-    public Optional<Group> createGroup(String name, String description, GroupType type, Set<String> requiredSkills) {
-        return groupStore.get()
-                .create(name, description, type, requiredSkills);
-    }
-
-
-    /**
-     * Update description for a group
-     *
-     * @param groupId        Group ID for group
-     * @param description    New description
-     * @param type           Type of group
-     * @param requiredSkills Set of skills needed to get automatically assigned to this group
-     * @return Updated group data
-     */
-    public Optional<Group> updateGroup(
-            final String groupId,
-            final String description,
-            final GroupType type,
-            final Set<String> requiredSkills) {
-        return groupStore.get()
-                .update(groupId,
-                        group -> group.withDescription(description)
-                                .withType(type)
-                                .withRequiredSkills(requiredSkills));
-    }
-
-    /**
-     * Delete an existing {@link Group}
-     *
-     * @param groupId ID for the existing group
-     * @return The updated group with {@link Group#isDeleted()} set to true
-     */
-    public boolean deleteGroup(String groupId) {
-        return groupStore.get().delete(groupId);
-    }
-
-    /**
-     * Read a group.
-     *
-     * @param groupId ID for the group to be read
-     * @return group or empty
-     */
-    public Optional<Group> readGroup(final String groupId) {
-        return groupStore.get().read(groupId);
-    }
-
-    /**
-     * Return all groups
-     *
-     * @return a list of groups
-     */
-    public List<Group> listGroups() {
-        return groupStore.get().list();
-    }
-
-    /**
-     * Add an {@link User} to a {@link Group}. A user can belong to multiple groups.
-     *
-     * @param groupId ID for an existing group
-     * @param userId  ID for an existing user
-     * @return Status of the operation
-     */
-    public boolean addUserToGroup(String groupId, String userId) {
-        val status = groupStore.get()
-                .addUserToGroup(groupId, userId);
-        log.info("User {} addition to group {}. Status: {}", groupId, userId, status);
-        return status;
-    }
-
-    /**
-     * Remove an user from a group
-     *
-     * @param groupId Group to be removed from
-     * @param userId  USer to be removed from group
-     * @return Status of operation
-     */
-    public boolean removeUserFromGroup(String groupId, String userId) {
-        val status = groupStore.get()
-                .removeUserFromGroup(groupId, userId);
-        log.info("User {} removal from group {}. Status: {}", groupId, userId, status);
-        return status;
-    }
-
-
-    /**
-     * Return a paginated list of users in the group
-     *
-     * @param groupId ID for an existing group
-     * @param start   Start offset of the page
-     * @param limit   Number oif records to return
-     * @return List of users in the group
-     */
-    public List<UserSummary> findUsersForGroup(String groupId, int start, int limit) {
-        val groupStoreImpl = this.groupStore.get();
-        val group = groupStoreImpl.read(groupId).orElse(null);
-        if (null == group || group.isDeleted()) {
-            return List.of();
-        }
-        val userIds = groupStoreImpl.findUsersForGroup(groupId, start, limit);
-        return userStore.get().getByIds(userIds);
-    }
-
-    public Optional<SkillDefinition> createSkill(final String name) {
-        return skillStore.get()
-                .createSkillDefinition(name);
-    }
-
-    public Optional<SkillDefinition> updateSkillDefinition(
-            final String skillId,
-            final String name) {
-        return skillStore.get()
-                .updateSkillDefinition(skillId, name);
-    }
-
-    public boolean deleteSkillDefinition(final String skillId) {
-        return skillStore.get()
-                .deleteSkillDefinition(skillId);
-    }
-
-    public Optional<SkillDefinition> getSkill(final String skillId) {
-        return skillStore.get()
-                .readSkillDefinition(skillId);
-    }
-
-    public Optional<SkillDefinition> addSkillValue(
-            final String skillId,
-            final String value) {
-        return skillStore.get()
-                .addValueToSkillDefinition(skillId, value);
-    }
-
-    public Optional<SkillDefinition> removeSkillValue(
-            final String skillId,
-            final String valueId) {
-        return skillStore.get()
-                .removeValueFromSkillDefinition(skillId, valueId);
-    }
-
-    public boolean addUserSkill(String userId, Skill skill) {
-        return skillStore.get().associateSkillWithUser(userId, skill.getSkillId(), skill.getSkillValueId());
-    }
-
-    public boolean removeUserSkill(String userId, Skill skill) {
-        return skillStore.get().disassociateSkillWithUser(userId, skill.getSkillId(), skill.getSkillValueId());
-    }
-
-    public List<SkillDefinition> listSkillDefinitions() {
-        return skillStore.get().listSkillDefinitions();
-    }
-
-    public List<SkillValue> listSkillValues() {
-        return skillStore.get().listSkillValues();
-    }
-
-    public List<SkillValue> getSkillsForUser(String userId) {
-        return skillStore.get().listSkillsForUser(userId);
-    }
-
-    private void createToken(UserSummary user) {
-        val token = userActivationTokenStore.get()
-                .generate(user.getId(), new Date(new Date().getTime() + DEFAULT_TOKEN_VALIDITY_MS))
-                .orElse(null);
-        if (null == token) {
-            log.error("No token generated for user: {}", user);
-        }
-        else {
-            mailSender.sendAsync(new MailSender.Mail(List.of(user.getEmail()),
-                                                     "Activate your Conductor Account",
-                                                     "Activation Token: " + token.getToken(),
-                                                     List.of()));
-            log.info("Token for user: {} is [{}]", user, token);
-        }
-        //TODO::SEND EVENT TO BUS
-    }
-
-
-    private static String hash(String s) {
-        return BCrypt.withDefaults().hashToString(12, s.toCharArray());
-    }
-
-
-}
